requirements-frozen.txt
aiofiles==24.1.0
boto3==1.35.93
botocore==1.35.93
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.1
colorama==0.4.6
contourpy==1.3.1
convopyro @ git+https://github.com/Ripeey/Conversation-Pyrogram@a39cc9e0c1a689c9904dd5be5419246868daecfb
coverage==7.6.9
cycler==0.12.1
fonttools==4.55.3
google-auth==2.37.0
google-auth-oauthlib==1.2.1
gspread==6.1.4
idna==3.10
iniconfig==2.0.0
jmespath==1.0.1
kiwisolver==1.4.7
matplotlib==3.10.0
mysql-connector-python==9.1.0
numpy==2.2.1
oauthlib==3.2.2
packaging==24.2
pandas==2.2.3
pendulum==3.0.0
pillow==11.0.0
pluggy==1.5.0
pyaes==1.6.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyparsing==3.2.0
Pyrogram @ https://github.com/KurimuzonAkuma/pyrogram/archive/refs/tags/v2.1.21.zip#sha256=eff8150622c0e8c9d615b33853a4fc11d57ca793418c4a0da164cdad410d832c
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
s3transfer==0.10.4
six==1.17.0
tabulate==0.9.0
TgCrypto==1.2.5
time-machine==2.16.0
tzdata==2024.2
urllib3==2.3.0
